cli/cmd: move all CLI independent code into cluster subpackage
This draws clear distinction between the CLI code sitting in cli/cmd
package right now and Lokomotive "core" logic.

cli/cmd/cluster package still contains some UI parts, like user input
confirmation or output printed directly to standard output, which should
also be addressed in the future.

Refs #603
